requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.5.1
attrs==19.3.0
black==20.8b1
celery==4.3.0
certifi==2020.6.20
click==7.1.2
coverage==5.2
Django==3.0.8
django-extensions==3.0.2
django-redis==4.11.0
djangorestframework==3.11.0
flake8==3.8.3
fluent-logger==0.9.6
freezegun==0.3.15
gunicorn==20.0.4
isort==5.0.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
msgpack==0.6.2
packaging==20.4
pathspec==0.8.0
Pillow
pluggy==0.13.1
psutil==5.7.0
psycopg2-binary==2.8.5
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.7.2
pylint-django==2.0.15
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-black==0.3.10
pytest-cov==2.10.0
pytest-django==3.8.0
pytest-flake8==1.0.6
pytest-freezegun==0.4.1
pytest-isort==1.1.0
python-dateutil==2.8.1
pytz==2020.1
redis==3.5.3
regex==2020.6.8
sentry-sdk==0.16.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.5
wrapt==1.12.1
